66 view service part delete service part

+++ src/main/java/seedu/address/logic/commands/DeleteVehicleCommand.java
@@ -4,7 +4,6 @@ import static java.util.Objects.requireNonNull;
 
 import java.util.List;
 
-import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.idgen.IdGenerator;
 import seedu.address.model.Model;
@@ -24,10 +23,14 @@ public class DeleteVehicleCommand extends RedoableCommand {
 
     public static final String MESSAGE_DELETE_VEHICLE_SUCCESS = "Deleted Vehicle: %1$s";
 
-    private final Index targetIndex;
+    public static final String MESSAGE_VEHICLE_NOT_FOUND = "Vehicle not in AutoM8";
 
-    public DeleteVehicleCommand(Index targetIndex) {
-        this.targetIndex = targetIndex;
+    private static final Vehicle VEHICLE_DOES_NOT_EXIST = null;
+
+    private final int id;
+
+    public DeleteVehicleCommand(int id) {
+        this.id = id;
     }
 
     @Override
@@ -35,8 +38,14 @@ public class DeleteVehicleCommand extends RedoableCommand {
         requireNonNull(model);
         List<Vehicle> lastShownList = model.getFilteredVehicleList();
 
-        Vehicle vehicleToDelete = lastShownList.get(targetIndex.getZeroBased());
-        int id = vehicleToDelete.getId();
+        // Get vehicle based on given id
+        Vehicle vehicleToDelete = lastShownList.stream().filter(vehicle->
+                id == vehicle.getId()).findFirst().orElse(VEHICLE_DOES_NOT_EXIST);
+
+        if (vehicleToDelete == VEHICLE_DOES_NOT_EXIST) {
+            throw new CommandException(MESSAGE_VEHICLE_NOT_FOUND);
+        }
+
         model.deleteVehicle(vehicleToDelete);
         IdGenerator.setVehicleIdUnused(id);
         return new CommandResult(String.format(MESSAGE_DELETE_VEHICLE_SUCCESS, vehicleToDelete));
@@ -46,6 +55,6 @@ public class DeleteVehicleCommand extends RedoableCommand {
     public boolean equals(Object other) {
         return other == this // short circuit if same object
                 || (other instanceof DeleteVehicleCommand // instanceof handles nulls
-                && targetIndex.equals(((DeleteVehicleCommand) other).targetIndex)); // state check
+                && this.id == ((DeleteVehicleCommand) other).id); // state check
     }
 }
